db: move stats methods
